+++ src/node/server/serverPluginEsbuild.ts
@@ -2,7 +2,7 @@ import { Plugin } from '.'
 import { tjsxRE, transform } from '../esbuildService'
 import { readBody, genSourceMapString } from '../utils'
 
-export const esbuildPlugin: Plugin = ({ app, watcher, jsxConfig }) => {
+export const esbuildPlugin: Plugin = ({ app, jsxConfig }) => {
   app.use(async (ctx, next) => {
     await next()
     if (ctx.body && tjsxRE.test(ctx.path)) {
@@ -14,12 +14,6 @@ export const esbuildPlugin: Plugin = ({ app, watcher, jsxConfig }) => {
         res += genSourceMapString(map)
       }
       ctx.body = res
-    }
-  })
-
-  watcher.on('change', (file) => {
-    if (tjsxRE.test(file)) {
-      watcher.handleJSReload(file)
     }
   })
 }