+++ src/node/server/serverPluginHmr.ts
@@ -35,7 +35,7 @@ import slash from 'slash'
 import chalk from 'chalk'
 import hash_sum from 'hash-sum'
 import { SFCBlock } from '@vue/compiler-sfc'
-import { parseSFC, vueCache } from './serverPluginVue'
+import { parseSFC, vueCache, srcImportMap } from './serverPluginVue'
 import { cachedRead } from '../utils'
 import { FSWatcher } from 'chokidar'
 import MagicString from 'magic-string'
@@ -124,7 +124,9 @@ export const hmrPlugin: Plugin = ({ root, app, server, watcher, resolver }) => {
     const timestamp = Date.now()
     if (file.endsWith('.vue')) {
       handleVueReload(file, timestamp)
-    } else if (file.endsWith('.js')) {
+    } else if (!file.endsWith('.css') || file.endsWith('.module.css')) {
+      // everything except plain .css are considered HMR dependencies.
+      // plain css has its own HMR logic in ./serverPluginCss.ts.
       handleJSReload(file, timestamp)
     }
   })
@@ -149,6 +151,7 @@ export const hmrPlugin: Plugin = ({ root, app, server, watcher, resolver }) => {
     const prevDescriptor = cacheEntry && cacheEntry.descriptor
     if (!prevDescriptor) {
       // the file has never been accessed yet
+      debugHmr(`no existing descriptor found for ${file}`)
       return
     }
 
@@ -222,7 +225,13 @@ export const hmrPlugin: Plugin = ({ root, app, server, watcher, resolver }) => {
   }
 
   function handleJSReload(filePath: string, timestamp: number = Date.now()) {
-    // normal js file
+    // normal js file, but could be compiled from anything.
+    // bust the vue cache in case this is a src imported file
+    if (srcImportMap.has(filePath)) {
+      debugHmr(`busting Vue cache for ${filePath}`)
+      vueCache.del(filePath)
+    }
+
     const publicPath = resolver.fileToRequest(filePath)
     const importers = importerMap.get(publicPath)
     if (importers) {